+++ R/XcmsExperiment-functions.R
@@ -1147,29 +1147,29 @@ XcmsExperiment <- function() {
 #' @author Johannes Rainer
 #'
 #' @noRd
-.xcms_experiment_to_xcms_n_exp <- function(x) {
+.xcms_experiment_to_xcms_n_exp <- function(from) {
     requireNamespace("MSnbase", quietly = TRUE)
     n <- new("XCMSnExp")
-    if (!length(x))
+    if (!length(from))
         return(n)
     m <- new("MsFeatureData")
     ## works only if we have a MsBackendMzR
-    if (!inherits(spectra(x)@backend, "MsBackendMzR"))
-        stop("Can not convert 'x' to a 'XCMSnExp' object: 'spectra(x)' uses an",
-             " MS backend other than 'MsBackendMzR'. Currently, only coercion",
-             " of an object with a 'MsBackendMzR' is supported.")
+    if (!inherits(spectra(from)@backend, "MsBackendMzR"))
+        stop("Can not convert 'from' to a 'XCMSnExp' object: 'spectra(x)' uses",
+             "an MS backend other than 'MsBackendMzR'. Currently, only",
+             " coercion of an object with a 'MsBackendMzR' is supported.")
     ## works only if we have an empty processing queue
-    if (length(spectra(x)@processingQueue) > 0)
-        stop("Can not convert 'x' to a 'XCMSnExp' object: processing queue of ",
-             "the Spectra object is not empty.")
+    if (length(spectra(from)@processingQueue) > 0)
+        stop("Can not convert 'from' to a 'XCMSnExp' object: processing queue",
+             " of the Spectra object is not empty.")
     ## -> OnDiskMSnExp
     n@processingData <- new("MSnProcess",
                             processing = paste0("Data converted [", date(), "]"),
-                            files = fileNames(x),
+                            files = fileNames(from),
                             smoothed = NA)
-    n@phenoData <- new("NAnnotatedDataFrame", as.data.frame(sampleData(x)))
-    fd <- as.data.frame(spectra(x)@backend@spectraData)
-    fd$fileIdx <- match(fd$dataStorage, fileNames(x))
+    n@phenoData <- new("NAnnotatedDataFrame", as.data.frame(sampleData(from)))
+    fd <- as.data.frame(spectra(from)@backend@spectraData)
+    fd$fileIdx <- match(fd$dataStorage, fileNames(from))
     fd <- fd[, !colnames(fd) %in% c("dataStorage", "dataOrigin")]
     colnames(fd) <- sub("scanIndex", "spIdx", colnames(fd))
     colnames(fd) <- sub("rtime", "retentionTime", colnames(fd))
@@ -1180,7 +1180,7 @@ XcmsExperiment <- function() {
                                   fd$fileIdx, fd$spIdx,
                                   max(fd$fileIdx), max(fd$spIdx))
     n@featureData <- new("AnnotatedDataFrame", fd)
-    nf <- length(fileNames(x))
+    nf <- length(fileNames(from))
     n@experimentData <- new("MIAPE",
                             instrumentManufacturer = rep(NA_character_, nf),
                             instrumentModel = rep(NA_character_, nf),
@@ -1188,20 +1188,20 @@ XcmsExperiment <- function() {
                             analyser = rep(NA_character_, nf),
                             detectorType = rep(NA_character_, nf))
     ## -> XCMSnExp
-    if (hasChromPeaks(x)) {
-        chromPeaks(m) <- chromPeaks(x)
-        chromPeakData(m) <- chromPeakData(x)
+    if (hasChromPeaks(from)) {
+        chromPeaks(m) <- chromPeaks(from)
+        chromPeakData(m) <- chromPeakData(from)
     }
-    if (hasAdjustedRtime(x)) {
+    if (hasAdjustedRtime(from)) {
         art <- fd$retentionTime_adjusted
         names(art) <- rownames(fd)
         adjustedRtime(m) <- split(art, fd$fileIdx)
     }
-    if (hasFeatures(x))
-        featureDefinitions(m) <- DataFrame(featureDefinitions(x))
+    if (hasFeatures(from))
+        featureDefinitions(m) <- DataFrame(featureDefinitions(from))
     lockEnvironment(m, bindings = TRUE)
     n@msFeatureData <- m
-    n@.processHistory <- x@processHistory
+    n@.processHistory <- from@processHistory
     validObject(n)
     n
 }