+++ R/functions-OnDiskMSnExp.R
@@ -722,7 +722,7 @@ setReplaceMethod("dirname", "OnDiskMSnExp", function(path, value) {
 #' par(mfrow = c(1, 2))
 #' plot(res, precursorIntensity(tmt))
 #' plot(res_2, precursorIntensity(tmt))
-.estimate_prec_intensity <- function(x, ppm = 10,
+.estimate_prec_intensity <- function(x, ppm = 10, tolerance = 0,
                                      method = c("previous", "interpolation")) {
     method <- match.arg(method)
     pmz <- precursorMz(x)
@@ -739,7 +739,8 @@ setReplaceMethod("dirname", "OnDiskMSnExp", function(path, value) {
             before_int <- numeric()
             if (length(before_idx)) {
                 sp <- sps[[before_idx[length(before_idx)]]]
-                before_idx <- closest(pmz[i], sp@mz, ppm = ppm, tolerance = 0,
+                before_idx <- closest(pmz[i], sp@mz, ppm = ppm,
+                                      tolerance = tolerance,
                                       duplicates = "closest")
                 if (!is.na(before_idx)) {
                     before_rt <- sp@rt
@@ -792,48 +793,6 @@ setReplaceMethod("dirname", "OnDiskMSnExp", function(path, value) {
         }
     }
     pmi
-}
-
-#' @title Estimate precursor intensity for MS level 2 spectra
-#'
-#' @description
-#'
-#' `estimatePrecursorIntensity` determines the precursor intensity for a MS 2
-#' spectrum based on the intensity of the respective signal from the
-#' neighboring MS 1 spectra (i.e. based on the peak with the m/z matching the
-#' precursor m/z of the MS 2 spectrum). Based on parameter `method` either the
-#' intensity of the peak from the previous MS 1 scan is used
-#' (`method = "previous"`) or an interpolation between the intensity from the
-#' previous and subsequent MS1 scan is used (`method = "interpolation"`, which
-#' considers also the retention times of the two MS1 scans and the retention
-#' time of the MS2 spectrum).
-#'
-#' @param x `OnDiskMSnExp` or `XCMSnExp` object.
-#'
-#' @param ppm `numeric(1)` defining the maximal acceptable difference (in ppm)
-#'     of the precursor m/z and the m/z of the corresponding peak in the MS 1
-#'     scan.
-#'
-#' @param method `character(1)` defining the method how the precursor intensity
-#'     should be determined (see description above for details). Defaults to
-#'     `method = "previous"`.
-#'
-#' @param BPPARAM parallel processing setup. See [bpparam()] for details.
-#'
-#' @return `numeric` with length equal to the number of spectra in `x`. `NA` is
-#'     returned for MS 1 spectra or if no matching peak in a MS 1 scan can be
-#'     found for an MS 2 spectrum
-#'
-#' @author Johannes Rainer
-#'
-#' @md
-estimatePrecursorIntensity <- function(x, ppm = 10,
-                                       method = c("previous", "interpolation"),
-                                       BPPARAM = bpparam()) {
-    method <- match.arg(method)
-    unlist(bplapply(.split_by_file2(x, subsetFeatureData = FALSE),
-                    .estimate_prec_intensity, ppm = ppm, method = method,
-                    BPPARAM = BPPARAM), use.names = FALSE)
 }
 
 #' Helper function to convert an OnDiskMSnExp to a Spectra object. This will